+++ src/formatters/formatterOptions.ts
@@ -1,10 +1,18 @@
 export interface FormatterOptions {
-  /** Command arguments. {@link getFormatterArgs} are always appended to these. */
+  /**
+   * Command arguments. {@link getFormatterArgs} are always appended to these.
+   */
   args?: string[];
-  /** Command. If omitted, {@link getFormatterCommand} is used. */
+  /**
+   * Command. If omitted, {@link getFormatterCommand} is used.
+   */
   cmd?: string;
-  /** Command working dir. If omitted, the workspace or parent dir of the scope URI is used. */
+  /**
+   * Command working dir. If omitted, the workspace or parent folder of document is used.
+   */
   cwd?: string;
-  /** Process env. */
+  /**
+   * Process env
+   */
   env?: NodeJS.ProcessEnv;
 }